src/machine_assembly_robot/requirements.txt
Flask==3.1.0
kafka-python==2.1.5
psycopg2-binary==2.9.10